pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-migrate>=4.1.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "python-telegram-bot==13.15",
    "qrcode>=8.0",
    "requests>=2.32.3",
    "telegram>=0.0.1",
    "sqlalchemy>=2.0.39",
    "pillow>=11.1.0",
]
